+++ app/src/main/java/com/cmput301w23t09/qrhunter/profile/ProfileController.java
@@ -10,7 +10,6 @@ import android.widget.Toast;
 import com.cmput301w23t09.qrhunter.DatabaseChangeListener;
 import com.cmput301w23t09.qrhunter.GameController;
 import com.cmput301w23t09.qrhunter.R;
-import com.cmput301w23t09.qrhunter.player.Player;
 import com.cmput301w23t09.qrhunter.player.PlayerDatabase;
 import com.cmput301w23t09.qrhunter.qrcode.DeleteQRCodeFragment;
 import com.cmput301w23t09.qrhunter.qrcode.QRCode;
@@ -18,7 +17,6 @@ import com.cmput301w23t09.qrhunter.qrcode.QRCodeAdapter;
 import com.cmput301w23t09.qrhunter.qrcode.QRCodeDatabase;
 import com.cmput301w23t09.qrhunter.qrcode.QRCodeFragment;
 import com.cmput301w23t09.qrhunter.qrcode.ScoreComparator;
-import com.cmput301w23t09.qrhunter.util.DeviceUtils;
 import com.google.android.gms.tasks.OnSuccessListener;
 import com.google.firebase.firestore.CollectionReference;
 import com.google.firebase.firestore.FirebaseFirestore;
@@ -32,17 +30,17 @@ import java.util.Objects;
 import java.util.UUID;
 
 /** This is the controller for the profile fragment of the app */
-public class ProfileController implements DatabaseChangeListener {
+public abstract class ProfileController implements DatabaseChangeListener {
   /** This is the game controller that controls the content on screen. */
-  private final GameController gameController;
+  protected final GameController gameController;
   /** This is the profile fragment the controller handles */
-  private final ProfileFragment fragment;
+  protected final ProfileFragment fragment;
   /** This is the array of QRCode objects that the fragment displays */
   private ArrayList<QRCode> qrCodes;
   /** This is the adapter for displaying the QRCode objects */
   private QRCodeAdapter qrCodeAdapter;
   /** Device UUID of the profile */
-  private final UUID deviceUUID;
+  protected final UUID deviceUUID;
   /** This is the gridview showing the qr codes of the player */
   private GridView qrCodeList;
   /** This is the view that shows the qr codes of the player */
@@ -180,29 +178,7 @@ public class ProfileController implements DatabaseChangeListener {
    * This handles the action to take when the contact info button is clicked. Either displaying the
    * contact information or rendering the edit details fragment.
    */
-  public void handleContactButtonClick() {
-    if (deviceUUID.equals(DeviceUtils.getDeviceUUID(gameController.getActivity()))) {
-      // Display edit settings fragment
-      ProfileSettingsFragment settingsFragment =
-          new ProfileSettingsFragment(gameController, deviceUUID);
-      gameController.setBody(settingsFragment);
-    } else {
-      // Display contact info popup
-      PlayerDatabase.getInstance()
-          .getPlayerByDeviceId(
-              deviceUUID,
-              task -> {
-                if (task.getException() != null) {
-                  showMsg(
-                      "An exception occurred while trying to load this player's contact details.");
-                  return;
-                }
-
-                Player player = task.getData();
-                fragment.displayContactInfo(player.getEmail(), player.getPhoneNo());
-              });
-    }
-  }
+  public abstract void handleContactButtonClick();
 
   /**
    * This updates the order of qr codes shown
@@ -266,7 +242,7 @@ public class ProfileController implements DatabaseChangeListener {
 
   /** This sets up the listener for real-time database changes */
   public void addUpdater() {
-    // QRCodeDatabase.getInstance().addListener(this);
+    QRCodeDatabase.getInstance().addListener(this);
   }
 
   /**
@@ -302,7 +278,7 @@ public class ProfileController implements DatabaseChangeListener {
    *
    * @param msg The message to display
    */
-  private void showMsg(String msg) {
+  protected void showMsg(String msg) {
     Toast.makeText(gameController.getActivity(), msg, Toast.LENGTH_SHORT).show();
   }
